Create PhoneNumber component to insert when needed
