+++ assets/js/blocks/cart-checkout/checkout/form/address-step.js
@@ -5,6 +5,7 @@ import { useMemo } from '@wordpress/element';
 import PropTypes from 'prop-types';
 import { useCheckoutAddress } from '@woocommerce/base-hooks';
 import { useShippingDataContext } from '@woocommerce/base-context';
+import { AddressForm } from '@woocommerce/base-components/cart-checkout';
 
 /**
  * Internal dependencies
@@ -12,6 +13,7 @@ import { useShippingDataContext } from '@woocommerce/base-context';
 import BillingFieldsStep from './billing-fields-step';
 import ContactFieldsStep from './contact-fields-step';
 import ShippingFieldsStep from './shipping-fields-step';
+import PhoneNumber from './phone-number';
 import './style.scss';
 
 const AddressStep = ( {
@@ -35,6 +37,7 @@ const AddressStep = ( {
 		showBillingFields,
 	} = useCheckoutAddress();
 	const { needsShipping } = useShippingDataContext();
+
 	const addressFieldsConfig = useMemo( () => {
 		return {
 			company: {
@@ -63,25 +66,44 @@ const AddressStep = ( {
 			/>
 			{ needsShipping && (
 				<ShippingFieldsStep
-					addressFieldsConfig={ addressFieldsConfig }
-					billingFields={ billingFields }
-					defaultAddressFields={ defaultAddressFields }
-					requirePhoneField={ requirePhoneField }
-					setPhone={ setPhone }
-					setShippingAsBilling={ setShippingAsBilling }
-					setShippingFields={ setShippingFields }
 					shippingAsBilling={ shippingAsBilling }
-					shippingFields={ shippingFields }
-					showPhoneField={ showPhoneField }
-				/>
+					setShippingAsBilling={ setShippingAsBilling }
+				>
+					<AddressForm
+						id="shipping"
+						type="shipping"
+						onChange={ setShippingFields }
+						values={ shippingFields }
+						fields={ Object.keys( defaultAddressFields ) }
+						fieldConfig={ addressFieldsConfig }
+					/>
+					{ showPhoneField && (
+						<PhoneNumber
+							isRequired={ requirePhoneField }
+							value={ billingFields.phone }
+							onChange={ setPhone }
+						/>
+					) }
+				</ShippingFieldsStep>
 			) }
 			{ showBillingFields && (
-				<BillingFieldsStep
-					addressFieldsConfig={ addressFieldsConfig }
-					billingFields={ billingFields }
-					defaultAddressFields={ defaultAddressFields }
-					setBillingFields={ setBillingFields }
-				/>
+				<BillingFieldsStep>
+					<AddressForm
+						id="billing"
+						type="billing"
+						onChange={ setBillingFields }
+						values={ billingFields }
+						fields={ Object.keys( defaultAddressFields ) }
+						fieldConfig={ addressFieldsConfig }
+					/>
+					{ showPhoneField && ! needsShipping && (
+						<PhoneNumber
+							isRequired={ requirePhoneField }
+							value={ billingFields.phone }
+							onChange={ setPhone }
+						/>
+					) }
+				</BillingFieldsStep>
 			) }
 		</>
 	);
